Initial code for some transaction work.
I managed to fit most of the implementation in a trait, so it is easy to mixin.  only works for SpecificNamespace at the moment.

Per record metadata serialization/deserialization.

Naive 2pc protocol tx code. works but mostly experimental/proof of concept.

TestTx is a simple program which tests some of the functionality.